The Great Localization
Book of Translatoris, Verse 3:
In the beginning, gr8pefish created the mod and the hardcoded strings.
And the mod was without formatting, and un-translatable; and translators were upon the threshold of anger. And the Spirit of TehNut moved upon the face of the code.
And TehNut said, Let there be localization: and there was localization.
And TehNut saw the localization, and it was good; and TehNut divided the hardcoded from the cleanliness.