Fixed Botania API support
Now supports only backpack in the inventory, not equipped ones.
Equipped backpacks would not be used by Botania API as it examines only
inventories. Call to equipped backpack could be triggered by presence of
another backpack in the inventory, but this is ugly hack and you can't
really prevent multiple calls to equipped backpack (e.g. for counting
available items).