Fix Item Dupe #234
This fix handles two issues that cause item duplication:

1) The view was lazy, it updated the model when the windows was closed,
rather than when changes occured. Because of this, items where
effectively duplicated as soon as the container was changed, and rolled
back to not duped when the view was closed, unless the backpack was
somehow removed before the save would happen.

2) Minecraft needs to know when the inventory is no longer accessible
to the player and should forcibly close the window, that's what
`canInteractWith()` is for. Its implementation is what prevents item
duplication in vanilla minecraft containers, and should be properly
implemented in modded containers.