Improve network syncing and stuff
Added `getNBTShareTag()` to `ItemBackpack`, making it so that the
inventory data is removed from the NBT sent in network packets.
This reduces both the size and amount of packets sent by the server,
since now it doesn't consider necessary to update the state of backpack
stack in clients when the internal inventory changes. This affects the
network performance of both the holder of the backpack and players
nearby, which previously received an `SPacketEntityEquipment` every
time the backpack held in any of the player's hand changed its
inventory.

Added a missing check before triggering the opening of the backpack's
GUI. Since it was being triggered for both client and server side, it
was opening twice on client side, once for the client side trigger and
once when receiving the message from server to open the GUI. This may
have caused desync issues with high latency players, as the player
could access the GUI before the server was ready to handle it. With
the previous change done to `getNBTShareTag()`, this issue became more
apparent as this caused a short flash of the GUI in a previous state
before the message from the server was handled with the proper GUI
state.