Rollback sync improvement due to bugs
Using `getNBTShareTag()` introduced bugs:

1) There are some cases where the backpack GUI would not stop the
player from clicking on the opened backpack when it was opened using
the open backpack key rather than right clicking it. This would
actually cause a desync issue between client and server, and instead of
duping, the backpack would be voided if the player attempts to throw it
on the world.

2) This one was rather obvious but not serious, there are desync issues
now while using the damage bar upgrade, caused by the fact that the
client doesn't know about the contents of the backpack until it is
opened.