Fix blocked slot client side desync
Storing the blocked stack and checking it against the stack in a slot
using `ItemStack.areItemStacksEqual()` to see if it should be blocked
caused weird desync client side issues when saving the inventory
contents eagerly was introduced to prevent item duping. What caused
this is that `ItemStack.EMPTY` is created with a stack size of 1 but
when synced through the network is sent with a stack size of 0, making
the `areItemStacksEqual()` check fail client side.